Extract NativeAdvertDetailsContainer for reusability
