+++ Sources/Components/NativeAdvert/NativeAdvertGridView.swift
@@ -28,7 +28,7 @@ public final class NativeAdvertGridView: UIView {
         return imageView
     }()
 
-    private lazy var detailsContainer = NativeDetailsContainer(withAutoLayout: true)
+    private lazy var detailsContainer = NativeAdvertDetailsContainer(withAutoLayout: true)
 
     private lazy var settingsButton: UIButton = {
         let button = CogWheelButton(alignment: .right, autoLayout: true)
@@ -50,24 +50,25 @@ public final class NativeAdvertGridView: UIView {
         imageView.topAnchor.constraint(equalTo: container.topAnchor),
         imageView.leadingAnchor.constraint(equalTo: container.leadingAnchor),
         imageView.heightAnchor.constraint(equalTo: imageView.widthAnchor, multiplier: 1 / imageAspectRatio),
-
-        detailsContainer.trailingAnchor.constraint(equalTo: container.trailingAnchor),
-        detailsContainer.bottomAnchor.constraint(equalTo: container.bottomAnchor),
     ]
 
     private lazy var compactConstraints: [NSLayoutConstraint] = [
         imageView.widthAnchor.constraint(equalTo: container.widthAnchor),
 
-        detailsContainer.topAnchor.constraint(equalTo: imageView.bottomAnchor),
+        detailsContainer.topAnchor.constraint(equalTo: imageView.bottomAnchor, constant: .mediumSpacing),
+        detailsContainer.bottomAnchor.constraint(equalTo: container.bottomAnchor, constant: -.mediumSpacing),
         detailsContainer.leadingAnchor.constraint(equalTo: container.leadingAnchor),
+        detailsContainer.trailingAnchor.constraint(equalTo: container.trailingAnchor)
     ]
 
     private lazy var regularConstraints: [NSLayoutConstraint] = [
         imageView.widthAnchor.constraint(equalTo: container.widthAnchor, multiplier: 0.5),
         imageView.bottomAnchor.constraint(equalTo: container.bottomAnchor),
 
-        detailsContainer.topAnchor.constraint(equalTo: container.topAnchor),
-        detailsContainer.leadingAnchor.constraint(equalTo: imageView.trailingAnchor),
+        detailsContainer.topAnchor.constraint(equalTo: container.topAnchor, constant: .mediumLargeSpacing),
+        detailsContainer.bottomAnchor.constraint(equalTo: container.bottomAnchor, constant: -.mediumLargeSpacing),
+        detailsContainer.leadingAnchor.constraint(equalTo: imageView.trailingAnchor, constant: .mediumLargeSpacing),
+        detailsContainer.trailingAnchor.constraint(equalTo: container.trailingAnchor, constant: -.mediumLargeSpacing)
     ]
 
     // MARK: - Init
@@ -117,9 +118,9 @@ public final class NativeAdvertGridView: UIView {
 
     private func setColors() {
         if traitCollection.horizontalSizeClass == .regular {
-            detailsContainer.backgroundColor = .bgTertiary
+            backgroundColor = .bgTertiary
         } else {
-            detailsContainer.backgroundColor = .bgPrimary
+            backgroundColor = .bgPrimary
         }
     }
 
@@ -135,158 +136,6 @@ public final class NativeAdvertGridView: UIView {
         super.traitCollectionDidChange(previousTraitCollection)
         setConstraints()
         setColors()
-    }
-
-}
-
-private class NativeDetailsContainer: UIView {
-
-    // MARK: - Private properties
-
-    private var logoSize: CGFloat = 50
-
-    private lazy var container = UIView(withAutoLayout: true)
-
-    private lazy var adRibbonContainer: UIView = {
-        let view = UIView(withAutoLayout: true)
-        view.layer.borderWidth = 1
-        view.layer.borderColor = .tableViewSeparator
-        view.layer.cornerRadius = .smallSpacing
-        return view
-    }()
-
-    private lazy var adRibbon: UILabel = {
-        let view = Label(style: .detail, withAutoLayout: true)
-        return view
-    }()
-
-    private lazy var companyLabel: UILabel = {
-        let view = Label(style: .detail, withAutoLayout: true)
-        view.textColor = .textAction
-        return view
-    }()
-
-    private lazy var titleLabel: UILabel = {
-        let view = Label(style: .body, withAutoLayout: true)
-        view.numberOfLines = 2
-        return view
-    }()
-
-    private lazy var logoView: UIImageView = {
-        let imageView = UIImageView(withAutoLayout: true)
-        imageView.contentMode = .scaleAspectFit
-        return imageView
-    }()
-
-    // MARK: - Constraints
-
-    private lazy var sharedConstraints: [NSLayoutConstraint] = [
-        adRibbonContainer.leadingAnchor.constraint(equalTo: container.leadingAnchor),
-        adRibbonContainer.topAnchor.constraint(equalTo: container.topAnchor),
-
-        adRibbon.topAnchor.constraint(equalTo: adRibbonContainer.topAnchor, constant: .verySmallSpacing),
-        adRibbon.leadingAnchor.constraint(equalTo: adRibbonContainer.leadingAnchor, constant: .smallSpacing),
-        adRibbon.bottomAnchor.constraint(equalTo: adRibbonContainer.bottomAnchor, constant: -.verySmallSpacing),
-        adRibbon.trailingAnchor.constraint(equalTo: adRibbonContainer.trailingAnchor, constant: -.smallSpacing),
-
-        companyLabel.centerYAnchor.constraint(equalTo: adRibbonContainer.centerYAnchor),
-        companyLabel.leadingAnchor.constraint(equalTo: adRibbonContainer.trailingAnchor, constant: .smallSpacing),
-
-        titleLabel.leadingAnchor.constraint(equalTo: container.leadingAnchor),
-        titleLabel.trailingAnchor.constraint(equalTo: logoView.leadingAnchor, constant: -.mediumSpacing),
-
-        logoView.widthAnchor.constraint(equalToConstant: logoSize),
-        logoView.heightAnchor.constraint(equalToConstant: logoSize),
-        logoView.bottomAnchor.constraint(equalTo: container.bottomAnchor),
-        logoView.trailingAnchor.constraint(equalTo: container.trailingAnchor)
-    ]
-
-    private lazy var compactConstraints: [NSLayoutConstraint] = [
-        heightAnchor.constraint(greaterThanOrEqualToConstant: logoSize + .mediumLargeSpacing),
-
-        titleLabel.topAnchor.constraint(equalTo: companyLabel.bottomAnchor, constant: .mediumSpacing),
-        titleLabel.bottomAnchor.constraint(equalTo: container.bottomAnchor),
-
-        container.topAnchor.constraint(equalTo: topAnchor, constant: .mediumSpacing),
-        container.bottomAnchor.constraint(equalTo: bottomAnchor, constant: -.mediumSpacing),
-        container.leadingAnchor.constraint(equalTo: leadingAnchor),
-        container.trailingAnchor.constraint(equalTo: trailingAnchor)
-    ]
-
-    private lazy var regularConstraints: [NSLayoutConstraint] = [
-        titleLabel.topAnchor.constraint(equalTo: companyLabel.bottomAnchor, constant: .mediumLargeSpacing),
-        titleLabel.bottomAnchor.constraint(lessThanOrEqualTo: container.bottomAnchor),
-
-        container.topAnchor.constraint(equalTo: topAnchor, constant: .mediumLargeSpacing),
-        container.bottomAnchor.constraint(equalTo: bottomAnchor, constant: -.mediumLargeSpacing),
-        container.leadingAnchor.constraint(equalTo: leadingAnchor, constant: .mediumLargeSpacing),
-        container.trailingAnchor.constraint(equalTo: trailingAnchor, constant: -.mediumLargeSpacing)
-    ]
-
-    // MARK: - Init
-
-    override init(frame: CGRect) {
-        super.init(frame: .zero)
-        setup()
-    }
-
-    required init?(coder: NSCoder) {
-        fatalError("init(coder:) has not been implemented")
-    }
-
-    // MARK: - Setup
-
-    private func setup() {
-        addSubview(container)
-
-        adRibbonContainer.addSubview(adRibbon)
-
-        container.addSubview(adRibbonContainer)
-        container.addSubview(companyLabel)
-        container.addSubview(titleLabel)
-        container.addSubview(logoView)
-
-        NSLayoutConstraint.activate(sharedConstraints)
-
-        setConstraints()
-        setFonts()
-    }
-
-    func configure(with model: NativeAdvertViewModel, andImageDelegate imageDelegate: NativeAdvertImageDelegate?) {
-        if let imageUrl = model.logoImageUrl {
-            imageDelegate?.nativeAdvertView(setImageWithURL: imageUrl, onImageView: logoView)
-        }
-
-        adRibbon.text = model.ribbonText
-        companyLabel.text = model.sponsoredBy
-        titleLabel.text = model.title
-    }
-
-    // MARK: - Overrides
-
-    override func traitCollectionDidChange(_ previousTraitCollection: UITraitCollection?) {
-        super.traitCollectionDidChange(previousTraitCollection)
-
-        setConstraints()
-        setFonts()
-    }
-
-    private func setConstraints() {
-        if traitCollection.horizontalSizeClass == .regular {
-            NSLayoutConstraint.deactivate(compactConstraints)
-            NSLayoutConstraint.activate(regularConstraints)
-        } else {
-            NSLayoutConstraint.deactivate(regularConstraints)
-            NSLayoutConstraint.activate(compactConstraints)
-        }
-    }
-
-    private func setFonts() {
-        if traitCollection.horizontalSizeClass == .regular {
-            titleLabel.font = .title2
-        } else {
-            titleLabel.font = .body
-        }
     }
 
 }
